requirements.txt
streamlit
streamlit_authenticator
snowflake-snowpark-python>=0.9.0
snowflake-connector-python>=2.8.0
matplotlib
yfinance
streamlit_calendar 
plotly
pandas 
numpy
datetime 
yfinance
polygon-api-client
